+++ connectd/connectd.c
@@ -1687,6 +1687,7 @@ static void connect_init(struct daemon *daemon, const u8 *msg)
 				    &tor_password,
 				    &daemon->timeout_secs,
 				    &daemon->websocket_helper,
+				    &daemon->message_padding,
 				    &daemon->dev_fast_gossip,
 				    &dev_disconnect,
 				    &daemon->dev_no_ping_timer,